+++ app/upgrades.go
@@ -12,6 +12,7 @@ import (
 	bech32ibctypes "github.com/osmosis-labs/bech32-ibc/x/bech32ibc/types"
 
 	"github.com/umee-network/umee/v3/app/upgradev3"
+	"github.com/umee-network/umee/v3/app/upgradev3_3"
 	leveragetypes "github.com/umee-network/umee/v3/x/leverage/types"
 	oracletypes "github.com/umee-network/umee/v3/x/oracle/types"
 )
@@ -36,10 +37,17 @@ func (app *UmeeApp) registerV3_1to3_3_Upgrade(_ upgradetypes.Plan) {
 		planName,
 		func(ctx sdk.Context, plan upgradetypes.Plan, fromVM module.VersionMap) (module.VersionMap, error) {
 			ctx.Logger().Info("Upgrade handler execution", "name", planName)
-			err := bankkeeper.NewMigrator(app.BankKeeper).Migrate3_V046_4_To_V046_5(ctx)
+			ctx.Logger().Info("Run v3.3 migrator")
+			err := upgradev3_3.Migrator(app.GovKeeper, app.interfaceRegistry)(ctx)
 			if err != nil {
 				return fromVM, err
 			}
+			ctx.Logger().Info("Run x/bank v0.46.5 migration")
+			err = bankkeeper.NewMigrator(app.BankKeeper).Migrate3_V046_4_To_V046_5(ctx)
+			if err != nil {
+				return fromVM, err
+			}
+			ctx.Logger().Info("Run module migrations")
 			return app.mm.RunMigrations(ctx, app.configurator, fromVM)
 		})
 }
@@ -51,6 +59,12 @@ func (app *UmeeApp) registerV3_2to3_3_Upgrade(_ upgradetypes.Plan) {
 		planName,
 		func(ctx sdk.Context, plan upgradetypes.Plan, fromVM module.VersionMap) (module.VersionMap, error) {
 			ctx.Logger().Info("Upgrade handler execution", "name", planName)
+			ctx.Logger().Info("Run v3.3 migrator")
+			err := upgradev3_3.Migrator(app.GovKeeper, app.interfaceRegistry)(ctx)
+			if err != nil {
+				return fromVM, err
+			}
+			ctx.Logger().Info("Run module migrations")
 			return app.mm.RunMigrations(ctx, app.configurator, fromVM)
 		})
 }
