+++ app/upgradev3_3/proposal_legacy.go
@@ -1,4 +1,4 @@
-package mv2
+package upgradev3_3
 
 import (
 	"encoding/json"
